Extract test parameter type
