+++ tests/integration/test_onprem_download.py
@@ -1,31 +1,15 @@
 import logging
 import shutil
 from pathlib import Path
-from typing import TypedDict
 
 import earthaccess
 import pytest
 from earthaccess import Auth, DataGranules, Store
 
+from .param import TestParam
 from .sample import get_sample_granules
 
 logger = logging.getLogger(__name__)
-
-
-class TestParam(TypedDict):
-    provider_name: str
-
-    # How many of the top collections we will test, e.g. top 3 collections
-    n_for_top_collections: int
-
-    # How many granules we will query
-    granules_count: int
-
-    # How many granules we will randomly select from the query
-    granules_sample_size: int
-
-    # The maximum allowed granule size; if larger we'll try to find another one
-    granules_max_size_mb: int
 
 
 daacs_list: list[TestParam] = [
